requirements.txt
annotated-types==0.7.0
anyio==4.5.2
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
filelock==3.16.1
flask==3.0.3
fsspec==2024.10.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
hyperframe==6.0.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
langchain==0.0.27
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==24.2
pillow==10.4.0
portalocker==2.10.1
protobuf==5.28.3
pydantic==1.10.9
pydantic-core==2.27.1
python-dotenv==1.0.1
pywin32==308
PyYAML==6.0.2
qdrant-client==1.12.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.6
Whoosh==2.7.4
zipp==3.20.2